# openapi/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/cities":{"post":{"operationId":"CiudadController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCiudadDto"}}}},"responses":{"201":{"description":"Crea una nueva ciudad","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ciudad"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudades"]},"get":{"operationId":"CiudadController_findAll","parameters":[],"responses":{"200":{"description":"Obtiene todas las ciudades","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ciudad"}}}}}},"tags":["Ciudades"]}},"/api/v1/cities/{id}":{"get":{"operationId":"CiudadController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Obtiene una ciudad por su ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ciudad"}}}},"404":{"description":"Ciudad no encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudades"]},"put":{"operationId":"CiudadController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCiudadDto"}}}},"responses":{"200":{"description":"Actualiza una ciudad por su ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ciudad"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Ciudad no encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudades"]},"delete":{"operationId":"CiudadController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Elimina una ciudad por su ID"},"404":{"description":"Ciudad no encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudades"]}},"/api/v1/cities/{cityId}/supermarkets/{supermarketId}":{"post":{"operationId":"CiudadSupermecadoController_addSupermarketToCity","parameters":[{"name":"cityId","required":true,"in":"path","schema":{"type":"string"}},{"name":"supermarketId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Supermercado añadido exitosamente a la ciudad","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ciudad"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"412":{"description":"Supermercado No existe","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudad - Supermercado"]},"get":{"operationId":"CiudadSupermecadoController_findSupermarketFromCity","parameters":[{"name":"cityId","required":true,"in":"path","schema":{"type":"string"}},{"name":"supermarketId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Obtiene un supermercado específico de la ciudad por su ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Supermercado"}}}},"404":{"description":"Supermercado o Ciudad no encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudad - Supermercado"]},"delete":{"operationId":"CiudadSupermecadoController_deleteSupermarketFromCity","parameters":[{"name":"cityId","required":true,"in":"path","schema":{"type":"string"}},{"name":"supermarketId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Supermercado eliminado exitosamente de la ciudad"},"404":{"description":"Supermercado no encontrado"}},"tags":["Ciudad - Supermercado"]}},"/api/v1/cities/{cityId}/supermarkets":{"get":{"operationId":"CiudadSupermecadoController_findSupermarketsFromCity","parameters":[{"name":"cityId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Obtiene todos los supermercados de la ciudad","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Supermercado"}}}}}},"tags":["Ciudad - Supermercado"]},"put":{"operationId":"CiudadSupermecadoController_updateSupermarketsFromCity","parameters":[{"name":"cityId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CiudadSupermercadoDto"}}}},"responses":{"200":{"description":"Actualiza los supermercados de la ciudad","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Supermercado"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Ciudad no encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"412":{"description":"Supermercado No existe","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Ciudad - Supermercado"]}},"/api/v1/supermarkets":{"post":{"operationId":"SupermercadoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupermercadoDto"}}}},"responses":{"201":{"description":"Crea un nuevo supermercado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Supermercado"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Supermercados"]},"get":{"operationId":"SupermercadoController_findAll","parameters":[],"responses":{"200":{"description":"Obtiene todos los supermercados","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Supermercado"}}}}}},"tags":["Supermercados"]}},"/api/v1/supermarkets/{id}":{"get":{"operationId":"SupermercadoController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Obtiene un supermercado por su ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Supermercado"}}}},"404":{"description":"Supermercado no encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Supermercados"]},"put":{"operationId":"SupermercadoController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupermercadoDto"}}}},"responses":{"200":{"description":"Actualiza un supermercado por su ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Supermercado"}}}},"400":{"description":"Datos de entrada inválidos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Supermercado no encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Supermercados"]},"delete":{"operationId":"SupermercadoController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Elimina un supermercado por su ID"},"404":{"description":"Supermercado no encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Supermercados"]}}},"info":{"title":"Supermarket API","description":"API para la consulta de la oferta de supermercados en las ciudades","version":"Version: 1","contact":{}},"tags":[{"name":"Ciudades","description":"Gestión de ciudades"},{"name":"Ciudad - Supermercado","description":"Gestión de supermercados por ciudad "},{"name":"Supermercados","description":"Gestión de supermercados"}],"servers":[{"url":"http://localhost:3000","description":"Local server"}],"components":{"schemas":{"CreateCiudadDto":{"type":"object","properties":{"nombre":{"type":"string","description":"Nombre de la ciudad","example":"Ciudad de la plata"},"pais":{"type":"string","description":"País al que pertenece la ciudad","example":"Argentina"},"numeroHabitantes":{"type":"number","description":"Número de habitantes de la ciudad","example":1000000}},"required":["nombre","pais","numeroHabitantes"]},"Ciudad":{"type":"object","properties":{"id":{"type":"number","example":1,"description":"Identificador único de la ciudad"},"nombre":{"type":"string","example":"Ciudad del Cairo","description":"Nombre de la ciudad"},"pais":{"type":"string","example":"Egipto","description":"País al que pertenece la ciudad"},"numeroHabitantes":{"type":"number","example":1000000,"description":"Número de habitantes de la ciudad"},"supermercados":{"description":"Lista de supermercados asociados a la ciudad","type":"array","items":{"$ref":"#/components/schemas/Supermercado"}}},"required":["id","nombre","pais","numeroHabitantes"]},"Supermercado":{"type":"object","properties":{"id":{"type":"number","example":1,"description":"Identificador único del supermercado"},"nombre":{"type":"string","example":"Supermercado Ejemplo","description":"Nombre del supermercado"},"longitud":{"type":"number","example":123.456789,"description":"Longitud geográfica del supermercado"},"latitud":{"type":"number","example":45.678901,"description":"Latitud geográfica del supermercado"},"paginaWeb":{"type":"string","example":"https://www.supermercadoejemplo.com","description":"Página web del supermercado"},"ciudades":{"description":"Lista de ciudades asociadas al supermercado","type":"array","items":{"$ref":"#/components/schemas/Ciudad"}}},"required":["id","nombre","longitud","latitud","paginaWeb"]},"ErrorResponseDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"Código de estado HTTP de la respuesta","example":404},"error":{"type":"string","description":"Error generado","example":"Not Found "},"message":{"type":"string","description":"Mensaje descriptivo del error","example":"Longitud del campo invalida / ID no encontrado"}},"required":["statusCode","error","message"]},"CiudadSupermercadoDto":{"type":"object","properties":{"supermercadoId":{"example":[{"id":"exampleId"}],"description":"Lista de identificadores de supermercados.","type":"array","items":{"type":"object"}}},"required":["supermercadoId"]},"CreateSupermercadoDto":{"type":"object","properties":{"nombre":{"type":"string","description":"Nombre del supermercado","example":"Supermercado Don Chucho"},"longitud":{"type":"number","description":"Longitud geográfica del supermercado","example":123.456789},"latitud":{"type":"number","description":"Latitud geográfica del supermercado","example":45.678901},"paginaWeb":{"type":"string","description":"Página web del supermercado","example":"https://www.supermercado-don-chucho.com"}},"required":["nombre","longitud","latitud","paginaWeb"]}}}}